// src/resources/enums.ts
export enum AuthSessionType {
  Fortnite = "fortnite",
  FortniteClientCredentials = "fortniteClientCredentials",
  EOS = "eos",
  Launcher = "launcher",
}

export enum AuthSessionStoreKey {
  Fortnite = "fortnite",
  FortniteClientCredentials = "fortniteClientCredentials",
  FortniteEOS = "fortniteEOS",
  Launcher = "launcher",
}

export enum SignedMessageType {
  Persistent = "Persistent",
  Party = "Party",
}

export enum ConversationType {
  DirectMessage = "DirectMessage",
  Party = "Party",
}