// src/resources/httpResponses.ts
export interface PublicKeyData {
  key: string;
  account_id: string;
  key_guid: string;
  kid: string;
  expiration: string;
  jwt: string;
  type: string;
}